+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 
+- Add extern function `free_array_len_cap` such vector pointers can be released and deprecated `free_array_len`
+  since it didn't account for capacity size.
 - Add `From` trait to convert `AccountKeys` into `AccountPublicKeys`.
 - Add `singleton` and `new` function to `AccountAccessStructure`.
 - Export `PublicKey`, `SecretKey`, and `Signature` type from `ed25519_dalek` crate.